desktop/requirements.txt
pillow
numpy
matplotlib
ttkbootstrap
pyinstaller
